\name{distrDefine}
\alias{distrDefine}
\alias{DgenericDistrDefine}
\alias{genericDistrDefine}
\title{
Definition of Random Variables
}
\description{In this window the user can define any random variable by providing its
parameters (for a well-known random variable), its distribution function, its density 
function (for a generic absolutely continuous random variable), or its mass probability
function (for a generic discrete random variable).
}

