\name{intervaloConfianzaMedia}
\alias{contrasteHipotesisMedia}
\alias{intervaloConfianzaMedia}
\title{
Confidence interval or hypothesis testing for the mean of a Normal variable
}
\description{In this graphical interface, the data selection is made to perform the calculation of the confidence interval or the hypothesis testing for the mean of a Normal variable.

This interface will call the statistical functions \code{\link{MKV.test}} and \code{\link{t.test}}, depending, respectively, on whether the population variance is known or not.
}
