\name{intervaloConfianzaMediasIndependientes}
\alias{intervaloConfianzaMediasIndependientes}
\alias{contrasteHipotesisMediasIndependientes}
\title{
Confidence interval or hypothesis testing for the difference in means of two independent Normal variables
}
\description{In this graphical interface, the data selection is made to perform the calculation of the confidence interval or the hypothesis testing for the difference in means of two independent Normal variables.

This interface will call the statistical functions \code{\link{DMKV.test}} and \code{\link{t.test}}, depending, respectively, on whether the population variances are known or not.
}
