\name{intervaloConfianzaVarianza}
\alias{contrasteHipotesisVarianza}
\alias{intervaloConfianzaVarianza}
\title{
Confidence interval or hypothesis testing for the Variance
}
\description{
In this graphical interface, the data selection is made to perform the calculation of the confidence interval or the hypothesis testing for the variance of a Normal variable.

This interface will call the statistical functions \code{\link{VKM.test}} and \code{\link{VUM.test}}, depending, respectively, on whether the population mean is known or not.
}
