\name{IndexNumbers}
\alias{Sindexnumbers}
\alias{BCindexnumbers}
\alias{Pindexnumbers}
\alias{Cindexnumbers}
\alias{Deflation}
\title{
Index Numbers menu
}
\description{In this menu the user can perform some calculations related to index numbers.

This menu will call the functions for calculating simple index numbers and making base changes (\code{\link{Sindex}}), for calculating complex index numbers (\code{\link{ComplexIN}}), for calculating price indices (\code{\link{priceIndexNum}}), and for deflation of economic series (\code{\link{Deflat}}).
}