\name{RandomANOVA}
\alias{oneWayAnovaRE}
\title{
One-Way ANOVA with random effects.
}
\description{In this menu the user can perform some calculations related to One-Way ANOVA with random effects.

This menu will call the functions for calculating the ANOVA table (\code{\link[stats]{aov}} from package \code{stats}) and the estimations of variance components using the maximum likelihood method and the REstricted Maximum Likelihood (REML) method (\code{\link[lme4]{lmer}} from package \code{lme4}).
}