\name{RcmdrPlugin.TeachStat-internal}
\title{Internal RcmdrPlugin.TeachStat objects}
\alias{resumenVariablesCualitativas}
\alias{resumenNumericoVDiscreta}
\alias{resumenNumericoPonderada}
\alias{resumenNumerico}
\alias{resumenDatosTabulados}
\alias{intervaloConfianzaProporcion}
\alias{intervaloConfianzaMediasPareadas}
\alias{intervaloConfianzaDiferenciaProporciones}
\alias{contrastHipotesisProporcion}
\alias{contrasteHipotesisMediasPareadas}
\alias{contrasteHipotesisDiferenciaProporciones}
\alias{intervaloConfianza2Varianzas}
\alias{contrasteHipotesis2Varianzas}
\alias{moda}
\alias{percentil}
\alias{pprint}
\alias{randomnessFTest}
\alias{randomnessNTest}
\alias{binVariable2}
\alias{newDataSet2}
\alias{numericToFactor2}
\alias{reorderFactors}
\alias{removeRedundantExtension}
\alias{variablesandtypesDataSet}
\description{Internal RcmdrPlugin.TeachStat objects}
\details{These are not to be called by the user.}
\keyword{internal}
