\name{RcmdrPlugin.TeachStat-package}
\alias{RcmdrPlugin.TeachStat-package}
\alias{RcmdrPlugin.TeachStat}
\docType{package}
\title{
RCommander plugin for teaching statistical methods. 
}
\description{
It adds a new menu for making easier the teaching of the main concepts about the main statistical methods.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.TeachStat\cr
Type: \tab Package\cr
Version: \tab 1.0.10-0\cr
Date: \tab 2018-12-11\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Tomás R. Cotos Yañez <cotos@uvigo.es>\cr
Manuel A. Mosquera Rodríguez <mamrguez@uvigo.es>\cr
Ana Pérez González <anapg@uvigo.es>\cr
Benigno Reguengo Lareo <benireguengo@gmail.com>
}

\keyword{ package }
\seealso{
\code{\link{Rcmdr}}.
}

