\name{RcmdrPlugin.TeachStat-internal}
\title{Internal RcmdrPlugin.TeachStat objects}
\alias{resumenVariablesCualitativas}
\alias{resumenNumericoVDiscreta}
\alias{resumenNumericoVContinua}
\alias{resumenNumericoPonderada}
\alias{resumenNumerico}
\alias{resumenDatosTabulados}
\alias{intervaloConfianzaProporcion}
\alias{intervaloConfianzaMediasPareadas}
\alias{intervaloConfianzaDiferenciaProporciones}
\alias{contrastHipotesisProporcion}
\alias{contrasteHipotesisMediasPareadas}
\alias{contrasteHipotesisDiferenciaProporciones}
\alias{intervaloConfianza2Varianzas}
\alias{contrasteHipotesis2Varianzas}
\alias{tabla.frec.cuantitativa}
\alias{moda}
\alias{percentil}
\alias{pprint}
\alias{randomnessFTest}
\alias{randomnessNTest}
\alias{binVariable2}
\alias{newDataSet2}
\alias{numericToFactor2}
\alias{reorderFactors}
\alias{removeRedundantExtension}
\alias{variablesandtypesDataSet}
\alias{col2hex}
\alias{rgb2col}
\alias{convert}
\alias{pickColor}
\alias{findNear}
\alias{FProbabilitiesInter}
\alias{PoissonProbabilitiesInter}
\alias{binomialProbabilitiesInter}
\alias{chisqProbabilitiesInter}
\alias{exponentialProbabilitiesInter}
\alias{geomProbabilitiesInter}
\alias{negbinomialProbabilitiesInter}
\alias{normalProbabilitiesInter}
\alias{tDistribution}
\alias{tProbabilitiesInter}
\alias{uniformProbabilitiesInter}
\alias{DgenericDistrCharac}
\alias{DgenericDistrMass}
\alias{DgenericDistrPlot}
\alias{DgenericDistrQuantiles}
\alias{DgenericDistrSamples}
\alias{DgenericProbabilitiesInter}
\alias{FDistrCharac}
\alias{FDistrPlot}
\alias{FDistrQuantiles}
\alias{FDistrSamples}
\alias{PoissonDistrCharac}
\alias{PoissonDistrMass}
\alias{PoissonDistrPlot}
\alias{PoissonDistrQuantiles}
\alias{PoissonDistrSamples}
\alias{binomialDistrCharac}
\alias{binomialDistrMass}
\alias{binomialDistrPlot}
\alias{binomialDistrQuantiles}
\alias{binomialDistrSamples}
\alias{chisqDistrCharac}
\alias{chisqDistrPlot}
\alias{chisqDistrQuantiles}
\alias{chisqDistrSamples}
\alias{exponentialDistrCharac}
\alias{exponentialDistrPlot}
\alias{exponentialDistrQuantiles}
\alias{exponentialDistrSamples}
\alias{genericDistrCharac}
\alias{genericDistrPlot}
\alias{genericDistrQuantiles}
\alias{genericDistrSamples}
\alias{genericProbabilitiesInter}
\alias{geomDistrCharac}
\alias{geomDistrMass}
\alias{geomDistrPlot}
\alias{geomDistrQuantiles}
\alias{geomDistrSamples}
\alias{negbinomialDistrCharac}
\alias{negbinomialDistrMass}
\alias{negbinomialDistrPlot}
\alias{negbinomialDistrQuantiles}
\alias{negbinomialDistrSamples}
\alias{normalDistrCharac}
\alias{normalDistrPlot}
\alias{normalDistrQuantiles}
\alias{normalDistrSamples}
\alias{tDistrCharac}
\alias{tDistrPlot}
\alias{tDistrQuantiles}
\alias{tDistrSamples}
\alias{uniformDistrCharac}
\alias{uniformDistrPlot}
\alias{uniformDistrQuantiles}
\alias{uniformDistrSamples}
\alias{checkNames}
\alias{isstring}
\description{Internal RcmdrPlugin.TeachStat objects}
\details{These are not to be called by the user.}
\keyword{internal}


