\name{twoOrMoreLevelFactors}
\alias{twoOrMoreLevelFactors}
\alias{twoOrMoreLevelFactorsP}
\title{
RcmdrPlugin.TeachStat Utility Functions
}
\description{
\code{twoOrMoreLevelFactorsP()} returns TRUE if there is at least one factor in the active dataset that has two or more levels.

\code{twoOrMoreLevelFactors()} returns the object name of those factors that are active in the dataset that have at least two levels.
}

