\name{RcmdrPlugin.TeachStat-package}
\alias{RcmdrPlugin.TeachStat-package}
\alias{RcmdrPlugin.TeachStat}
\docType{package}
\title{
R Commander plugin for teaching statistical methods. 
}
\description{
It adds a new menu for making easier the teaching of the main concepts about the main statistical methods.
}
\details{
\tabular{ll}{
Package: \tab \code{RcmdrPlugin.TeachStat}\cr
Type: \tab Package\cr
Version: \tab 1.1.2\cr
Date: \tab 2023-11-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Tomás R. Cotos Yañez <cotos@uvigo.gal>\cr
Manuel A. Mosquera Rodríguez <mamrguez@uvigo.gal>\cr
Ana Pérez González <anapg@uvigo.gal>\cr
Benigno Reguengo Lareo <benireguengo@gmail.com>
}

\keyword{ package }
\seealso{
\code{\link{Rcmdr}}.
}

